fix: Split 1.18 and 1.19 completely It's causing problems apparently
