+++ cobblegen-1.18.2-compat/src/main/resources/fabric.mod.json
@@ -20,6 +20,6 @@
     },
     "depends": {
         "fabricloader": "*",
-        "minecraft": ">=1.18.2 <=1.19.2"
+        "minecraft": "~1.18.2"
     }
 }
