+++ cobblegen-1.19-compat/build.gradle
@@ -1,10 +1,6 @@
 archivesBaseName += "-1.19-compat"
 
 dependencies {
-    // 1.18.2 and 1.19.x (except for 1.19.3 and 1.19.4) is still pretty similar,
-    // so instead of creating new CompatImpl we'll install extend 1.18's Compat
-    api project(path: ':cobblegen-1.18.2-compat', configuration: 'namedElements')
-
     minecraft "com.mojang:minecraft:${project.minecraft_version_1_19_2}"
     //noinspection GradlePackageVersionRange
     mappings "net.fabricmc:yarn:${project.yarn_mappings_1_19_2}:v2"
