+++ cobblegen-1.19.3-compat/src/main/java/io/github/null2264/cobblegen/CompatImpl.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class CompatImpl extends Compat
+public class CompatImpl implements Compat
 {
     @Override
     public MutableText translatableText(String string) {
